--- src/app/(main)/layout.tsx
@@ -1,24 +1,0 @@
-import MainDetail from "../../components/templates/main/MainDetail";
-import type { Metadata } from "next";
-import { Inter } from "next/font/google";
-import "../../styles/globals.css";
-
-export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
-};
-
-export default function RootLayout({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
-  return (
-    <html lang="en">
-      <body>
-        <MainDetail />
-        {children}
-      </body>
-    </html>
-  );
-}
