--- src/app/(main)/page.stories.tsx
@@ -1,8 +1,0 @@
-import Page from "./page";
-
-export default {
-  title: "app/page",
-  component: Page,
-};
-
-export const MainPage = () => <Page />;
