+++ src/app/(main)/test/page.tsx
@@ -1,0 +1,12 @@
+import MainDetail from "@/components/templates/main/MainDetail";
+import React from "react";
+
+const page = () => {
+  return (
+    <div>
+      <MainDetail />
+    </div>
+  );
+};
+
+export default page;
