+++ src/app/page.tsx
@@ -1,0 +1,10 @@
+import Main from "@/components/templates/main/Main";
+import Image from "next/image";
+
+export default function Home() {
+  return (
+    <div>
+      <Main />
+    </div>
+  );
+}
