+++ src/components/templates/main/MainDetail.tsx
@@ -1,14 +1,22 @@
 "use client";
 import DetailProfile from "../../atoms/profile/detailProfile/DetailProfile";
-import React from "react";
+import React, { useState } from "react";
 import test from "../../../../public/icons/testProfile.png";
 import Letter from "../../atoms/letter/Letter";
 import BottomFixedBtn from "../../atoms/button/bottomFixed/BottomFixed";
+import LetterModal from "@/components/organisms/modal/LetterModal";
 
 const MainDetail = () => {
+  const [isLetterModalOpened, setLetterModalOpened] = useState(false);
   return (
     <>
-      <div className="mt-20 relative  w-360 flex flex-col items-center justify-start">
+      {isLetterModalOpened && (
+        <LetterModal
+          isLetterModalOpened={isLetterModalOpened}
+          setIsLetterModalOpened={setLetterModalOpened}
+        />
+      )}
+      <div className="mt-20 relative tablet:min-h-full w-360 flex flex-col items-center justify-start">
         <div className="mb-28">
           <DetailProfile
             profileImg={test}
@@ -74,9 +82,12 @@ const MainDetail = () => {
             </div>
           </div>
         </div>
-        {/* <div className="m-16 absolute bottom-0 right-0">
+        <div
+          onClick={() => setLetterModalOpened(!isLetterModalOpened)}
+          className="m-16 fixed  bottom-0 right-0"
+        >
           <BottomFixedBtn usage={"write"} />
-        </div> */}
+        </div>
       </div>
     </>
   );
